% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Solver for a general problem of the form: min (loss + Orthopen)}
\usage{
main(X, Y, lambda = 1, step_size = 0.1, verbose = 0,
  stop_no_improve = 100, max_iter = 1e+06, K = NULL, disjoint = TRUE,
  logistic = FALSE, enet = FALSE)
}
\arguments{
\item{X}{\eqn{MxP} observations matrix (features are in columns)}

\item{Y}{\eqn{MxT} observed output matrix for T different tasks}

\item{lambda}{a regularization parameter (default \code{1})}

\item{step_size}{step for gradient descent (default \code{0.1})}

\item{verbose}{option (default \code{0})}

\item{stop_no_improve}{number of gradient descent steps without improvment before stopping (default: \code{100})}

\item{max_iter}{maximum number of iterations before stopping optimization (default: \code{1000000})}

\item{K}{\eqn{PxP} orthogonality constraints matrix (default: diagonal matrix --> no orthogonality constraint)}

\item{disjoint}{if \code{TRUE} add contraints for disjoint supports (default: TRUE)}

\item{logistic}{if \code{TRUE} change loss function to logistic loss (for classification problems)}

\item{enet}{if \code{TRUE}, add a \code{L1} penalization to the \code{L2} penalization, using elastic net formula (single parameter lambda, assuming \eqn{enet = 0.5*L2 + 0.5*L1})}
}
\value{
a list containing three elements\itemize{\item \code{W}: optimal \eqn{PxT} matrix for objective function minimum \item obj: objective function value at W \item imax: number of steps before reaching optimum
}
}
\description{
\code{main}
}
\examples{
#solve orthogonal columns problem 
# min_W 1/2 norm( X\%*\%W - Y )^2 + lambda ||W||_orthopen
NVAR=10
NTRAIN=100
T=3
K = matrix(1,nrow=T,ncol=T)
# Generate random data and random model
X <- matrix(rnorm(NTRAIN*NVAR),nrow=NTRAIN,ncol=NVAR)
# Random orthogonal matrix
W <- qr.Q(qr(matrix(rnorm(NVAR*T),nrow=NVAR,ncol=T)))
Y <- X \%*\% W + matrix(rnorm(NTRAIN*T),nrow=NTRAIN)
set.seed(42)
res <- main(X,Y,lambda = 0.1,K = K,disjoint = FALSE)
}
