% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{myproj}
\alias{myproj}
\title{Projection function for disjoint supports}
\usage{
myproj(w, v)
}
\arguments{
\item{w}{the weight matrix used in loss function}

\item{v}{the constraint matrix used for projection}
}
\description{
Projection function for disjoint supports
}
