% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection.R
\name{proj_disjoint}
\alias{proj_disjoint}
\title{Projection function for disjoint supports}
\usage{
proj_disjoint(w, v)
}
\arguments{
\item{\code{w}}{the weight matrix used in loss function}

\item{\code{v}}{the constraint matrix used for projection}
}
\value{
a list of the projected matrices \code{w} and \code{v}
}
\description{
Projection function for disjoint supports
}
